Add files via upload
In routine life students are not always able to get the perfect answer to their queries and questions raised, though they search on the internet, using a video or tutorial. When a student has any issue with their studies or academics, he/she would contact their faculty but if that faculty is unavailable, it creates havoc for the student to solve the problem. With the help of this website, students will post their queries and the faculty will respond with the answer and notes them, so it makes it easy for students to learn from there. Suppose, if a student is unable to understand the reply or the answers from the faculty, he can request a one-to-one meeting with the faculty that makes it easy for the student to learn the topic easily. Mostly all other learning platforms only provide tutorials and examples, but with this website, they will also be able to have a one-to-one meeting with their concerned mentor. In rural areas, students get the best faculty to learn things in a better way.